third_party/WebKit/LayoutTests/fast/frames/srcdoc/srcdoc-deep-nested-frames.html

<!DOCTYPE html>

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<iframe id="frame1" srcdoc="
    1 <iframe id='frame2' srcdoc='
        2 <iframe id=frame3 srcdoc=3></iframe>
'></iframe>
"></iframe>

<script>
document.getElementById('frame1').onload = function() {
    var frame2 = frame1.contentDocument.getElementById('frame2');
    var frame3 = frame2.contentDocument.getElementById('frame3');
    test(function() {
        assert_not_equals(frame3.contentDocument, null);
        assert_not_equals(frame3.clientHeight, 0);
    }, 'Nested srcdocs should not be blocked as self-referential frames');
};
</script>
